A cool little pong game.
Pong is a groundbreaking game that played a key role in popularizing arcade gaming. Developed by Allan Alcorn and published by Atari, it was the first commercially successful video game, laying the foundation for the modern video game industry.

The primary goal of this project was to gain a deeper understanding of how to effectively use SFML and C++ together in a cohesive and practical way.